Add lesson 4 homework files
This commit adds three files for lesson 4 homework. They include task.py, nontask.py, and a Readme.md description file. In task.py, various operations are demonstrated on python data structures including lists, tuples, sets and dictionaries. In nontask.py, a nested dictionary creation and operations on it are demonstrated. Feel free to skip it! The Readme.md file gives a detailed explanation of the contents and usage of both python script files.